requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
passlib[argon2]
python-jose[cryptography]
alembic
celery
redis
email-validator
psycopg2-binary
